Move self-tests from core.py to tests/
