+++ _stbt/core.py
@@ -3026,113 +3026,3 @@ def _hocr_elem_region(elem):
             extents = [int(x) for x in m.groups()]
             return Region.from_extents(*extents)
         elem = elem.getparent()
-
-# Tests
-# ===========================================================================
-
-
-def test_wait_for_motion_half_motion_str_2of4():
-    with _fake_frames_at_half_motion() as dut:
-        res = dut.wait_for_motion(consecutive_frames='2/4')
-        print res
-        assert res.time == 1466084606.
-
-
-def test_wait_for_motion_half_motion_str_2of3():
-    with _fake_frames_at_half_motion() as dut:
-        res = dut.wait_for_motion(consecutive_frames='2/3')
-        print res
-        assert res.time == 1466084606.
-
-
-def test_wait_for_motion_half_motion_str_4of10():
-    with _fake_frames_at_half_motion() as dut:
-        # Time is not affected by consecutive_frames parameter
-        res = dut.wait_for_motion(consecutive_frames='4/10', timeout_secs=20)
-        assert res.time == 1466084606.
-
-
-def test_wait_for_motion_half_motion_str_3of4():
-    with _fake_frames_at_half_motion() as dut:
-        try:
-            dut.wait_for_motion(consecutive_frames='3/4')
-            assert False, "wait_for_motion succeeded unexpectedly"
-        except MotionTimeout:
-            pass
-
-
-def test_wait_for_motion_half_motion_int():
-    with _fake_frames_at_half_motion() as dut:
-        try:
-            dut.wait_for_motion(consecutive_frames=2)
-            assert False, "wait_for_motion succeeded unexpectedly"
-        except MotionTimeout:
-            pass
-
-
-@contextmanager
-def _fake_frames_at_half_motion():
-    FRAMES = []
-    a = numpy.zeros((2, 2, 3), dtype=numpy.uint8)
-    b = numpy.ones((2, 2, 3), dtype=numpy.uint8) * 255
-
-    # Motion:                 v     v     v     v     v     v     v     v     v
-    data = [a, a, a, a, a, a, b, b, a, a, b, b, a, a, b, b, a, a, b, b, a, a, b]
-    #       ^                 ^
-    #       |                 L Motion starts here at timestamp 1466084606.
-    #       L Video starts here at timestamp 1466084600
-
-    FRAMES = [Frame(x, time=1466084600. + n) for n, x in enumerate(data)]
-
-    class FakeDisplay(object):
-        def get_frame(self, timeout_secs=10, since=0):  # pylint: disable=unused-argument
-            for f in FRAMES:
-                if f.time > since:
-                    return f
-            f = FRAMES[-1].copy()
-            f.time = since + 1
-            return f
-
-    class FakeTime(object):
-        def __init__(self, now):
-            self._now = now
-
-        def time(self):
-            return self._now
-
-        def sleep(self, duration):
-            self._now += duration
-
-    dut = DeviceUnderTest(display=FakeDisplay(), sink_pipeline=NoSinkPipeline(),
-                          _time=FakeTime(FRAMES[0].time))
-    yield dut
-
-
-def test_region_replace():
-    from nose.tools import raises
-
-    r = Region(x=10, y=20, width=20, height=30)
-
-    def t(kwargs, expected):
-        assert r.replace(**kwargs) == expected
-
-    @raises(ValueError)
-    def e(kwargs):
-        r.replace(**kwargs)
-
-    # No change
-    yield t, dict(x=10), r
-    yield t, dict(x=10, width=20), r
-    yield t, dict(x=10, right=30), r
-
-    # Not allowed
-    yield e, dict(x=1, width=2, right=3)
-    yield e, dict(y=1, height=2, bottom=3)
-
-    # Allowed  # pylint:disable=C0301
-    yield t, dict(x=11), Region(x=11, y=r.y, width=19, height=r.height)
-    yield t, dict(width=19), Region(x=10, y=r.y, width=19, height=r.height)
-    yield t, dict(right=29), Region(x=10, y=r.y, width=19, height=r.height)
-    yield t, dict(x=11, width=20), Region(x=11, y=r.y, width=20, height=r.height)
-    yield t, dict(x=11, right=21), Region(x=11, y=r.y, width=10, height=r.height)
-    yield t, dict(x=11, right=21, y=0, height=5), Region(x=11, y=0, width=10, height=5)
